+++ default_www/backend/modules/mailmotor/engine/helper.php
@@ -673,6 +673,45 @@ class BackendMailmotorCMHelper
 
 
 	/**
+	 * Returns all subscribers, regardless of the page limit CM gives us.
+	 *
+	 * @return	array
+	 * @param	string $listId	The list ID to get the subscribers from.
+	 */
+	public static function getSubscribers($listId)
+	{
+		// get list statistics, so we can obtain the total subscribers for this list
+		$listStats = self::getCM()->getListStatistics($listId);
+
+		// pagecount is calculated by getting the total amount of subscribers divided by 1k, which is the return limit for CM's getSubscribers()
+		$pageCount = (int) ceil($listStats['total_subscribers'] / 1000);
+
+		// reserve a result stack
+		$results = array();
+
+		// check if we have at least 1 page
+		if($listStats['total_subscribers'] !== 0)
+		{
+			// set the pagecount to 1 by default
+			$pageCount++;
+
+			// loop the total amount of pages and fetch the subscribers accordingly
+			for($i = $pageCount; $i != 0; $i--)
+			{
+				// get the subscribers
+				$subscribers = self::getCM()->getSubscribers($listId, null, $i, 1000);
+
+				// add the subscribers to the result stack
+				$results = array_merge($results, $subscribers);
+			}
+		}
+
+		// return the results
+		return $results;
+	}
+
+
+	/**
 	 * Returns the CampaignMonitor countries as pairs
 	 *
 	 * @return	array