+++ default_www/install/engine/step_2.php
@@ -51,7 +51,7 @@ class InstallerStep2 extends InstallerStep
 	public static function checkRequirements(array &$variables = null)
 	{
 		// define step
-		$step = (isset($_GET['step']) && in_array($_GET['step'], array('1', '2', '3', '4', '5', '6'))) ? (int) $_GET['step'] : 1;
+		$step = (isset($_GET['step']) && in_array($_GET['step'], array('1', '2', '3', '4', '5', '6', '7'))) ? (int) $_GET['step'] : 1;
 
 		// define constants
 		if(!defined('PATH_WWW') && !defined('PATH_LIBRARY')) self::defineConstants($step);