+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Bugfixes:
 * Bugfix: hover-event wasn't unbind correctly when sorting the widgets was done.
 * Bugfix: importing addresses into the mailmotor was borked, as reported on http://forkcms.lighthouseapp.com/projects/61890/tickets/143.
 * Bugfix: focusFirst was focusing on an element on hidden tabs, as reported on http://forkcms.lighthouseapp.com/projects/61890-fork-cms/tickets/153.
+* Bugfix: click on tab wasn't working decent in IE, as reported on http://forkcms.lighthouseapp.com/projects/61890/tickets/154. 
+* Bugfix: page-revisions were interfering with blog-revisions, as reported on http://forkcms.lighthouseapp.com/projects/61890/tickets/151.
 
 Improvements:
 * Core: when using datefields with till, from, range set, it will be validated according the type.
@@ -27,6 +29,7 @@ Improvements:
 * Core: content blocks can now be linked to a content block-template.
 * Core: locales analyse-tool will check only the active modules from now on.
 * Core: added a jQuery-plugin to implement a passwordGenerator.
+* Core: added the possibility to add attachments to the frontend/backend mailers.
 * Blog: creating categories can now be done without leaving the add/edit screen.
 * Blog: changes to improve the usability: no more default category, users are forced to select a category if there are multiple categories.
 * Blog: when filtered on a category and clicked on link to add a post the category will be prefilled.
@@ -37,10 +40,14 @@ Improvements:
 * Pages: when changing templates the textual-content isn't deleted anymore.
 * Locale: you can now import/export locale from/to xml. The installers also use xml's.
 * Locale: export for missing locale.
+* Locale: remove deprecated insertLocale function.
+* Locale: created an incredibly nasty hotfix for some deprecated PHP functionality.
 * Mailmotor: added extra validation (reported on http://forkcms.lighthouseapp.com/projects/61890/tickets/137).
 * Mailmotor: added extra validation for adding address(es).
 * Mailmotor: adding multiple addresses now uses the multipleTextbox-functionality.
 * Installer: refactored pages installation.
+* Installer: split up step languages & modules into 2 steps; moved db step behind those.
+* Installer: ask for backend interface languages seperate from frontend languages.
 
 2.1.0 (2011-03-14)
 --